Refactor standalone toggle button to separate file
Moved the RdsStandaloneToggleButton component to its own file (rds-standalone-toggle-button.tsx) and updated imports/exports in rds-toggle-button.tsx. This improves code organization and separation of concerns.
